Dynamic locations
This current version updates the possibility of creating dynamic locations of the email, their attachments (currently working only with excel attachments), the configuration, also generates already pre build structure of the connection and possible email white listing. Meaning upon executing index it will automatically generate necessary folders and fill in needed files. The reason why all of this is needed since if you use this project as a composer module ( library) you could add it to an existing project thus meaning it would only be handy writing the code if you could specify the locations of the folders... And also began making filter request in order to make option to use the filtered records for webservices...